+++ CHANGELOG.md
@@ -3,8 +3,13 @@
 ## master
 
 ## 1.0.0
+* Add page count and delivery ids
 * Disable donate and premium links
+* Enable global actions
 * Initial code release 
+* PDF counter for small ones
+* Relates docs to delivery record
+* Remove admin only restriction
 * Remove mass processor
 * Remove Process plugin action link
 * Set plugin URL to aihr.us